Add ARIA labels to improve accessibility and user experience
This commit adds ARIA labels to improve accessibility and user experience. The "title" attribute in the title tag in index.html has been replaced with an "aria-label" attribute to describe the website. The "alt" attributes have been added to the logo images in index.html and photographer.html, and the "alt" attributes of the media elements have been set to their respective titles. Additionally, ARIA labels have been added to the contact and lightbox buttons in photographer.html to provide more context to users. Finally, ARIA labels have also been added to the photographer's name and picture links in photographer.js.